+++ RELEASE_NOTES.md
@@ -1,12 +1,18 @@
 # NexusKey v2.1.9
 
+**✨ Cải tiến:**
+- Cải thiện thuật toán engine
+- Tham khảo thêm mã nguồn của Unikey để tối ưu engine, kiểm tra chính tả và chặn đặt dấu từ tiếng anh https://www.unikey.org/source.html. Cảm ơn bác Kim Long
+
 **🛠 Sửa lỗi:**
 - Sửa lỗi uu7 -> uư thay vì ưu
 - Sửa lỗi spell check không thoát dấu móc cho case e-u-w-w
-- Sửa lỗi smartswicth hoạt động không đúng (revert lại cách sử cũ)
 - Sửa lỗi kìaaaaaaaaaaaa
 - Sửa lỗi không hoàn tác dấu sau khi space VNI
 - Sửa lỗi smartswitch và exclude apps không hoạt động đúng trong một số trường hợp
+- Bỏ soft exclude
+- Sửa lỗi Ctrl + BS gây ghostkey
+- Sửa lỗi bỏ dấu sai cho các chữ kéo dài (kìaaaaaaaaa)
 
 ---
 Cảm ơn bạn đã chọn NexusKey! Mọi góp ý của bạn đều giúp bộ gõ ngày càng hoàn thiện hơn.